Basic Fire Scene
Basic components to make the fire mixing gameplay. Campfire with health, an enemy that goes to the campfire and deals damage to it, and the start of the mechanic to kill the enemy. Needs still different types of enemies, the mechanics to kill different types of enemies, enemy spawners, and ui. The connection to the title screen and the platforming section is also necessary