Fire Defense Scene Update
Fire defense system able to color mix now. All primary and secondary color enemies have been made and can be selectively killed depending on the fire selected. Working next on enemy spawners, campfire health system needs fixing, and ui. The connection to the title screen and the platforming section with inventory is also necessary but will probably not be finished in time lol.